+++ NEWS.md
@@ -6,6 +6,9 @@
 - Add support for toggling whether to search in hidden files (`M-s h` in the
   search prompt, or `urgrep-search-hidden-files` globally)
 
+### Breaking changes
+- `urgrep-run-command` now takes `directory` and `tool` as optional keys
+
 ---
 
 ## v0.1.1 (2023-06-07)